chore: improve running rails generator before test suite
running this in an initializer is "too late" in terms of how this would
actually be used. as such, it was hiding a bug. now instead of doing
this during application initialization, we're running before that. then
during combustion app initialization migrations will run. this allows us
to leave the schema.rb file blank.