fix: misc bug fixes from an integration test
this commit marks the first successful end-to-end test of the library!
WOOOOOOOOOOOOO! i successfully triggered a pipeline (from a rails
console) which created the necessary records, the job was picked up by a
worker, and executed successfully! the pipeline halted as the pipeline
advancer algorithm is not up-to-snuff yet so no further steps executed.