feat: introduce `pipeline.shutdown_timeout` configuration
we're moving to a concurrency model for the pipeline advancer.
currently, it only runs a single process tp update all configured
pipelines. this creates a massive bottleneck. in coming commits we'll
create a thread for each configured pipeline. i'm thinking for Pro we
fork a process for each configured pipeline and create a thread pool to
advance all pipelines for a specific klass. more on that later.