feat!: introduce concurrency model to pipeline advancer
this is step 1 of N for concurrencyifying the pipeline advancer
sub-system. originally, it was a single process that would attempt to
update ALL configured pipelines. this was a massive bottleneck. now, it
creates threads for each configured pipeline. once implementation is
finished, i'll run benchmarks to check the performance and re-design as
needed