feat: introduce logger level and source configurations
redoing logger config here. the `logger` is now not really recommended
to be used unless necessary (needing to log to a specific file). the
documentation will have specific instructions on what to do if needed.

anyway, now the logger FOR THE DUCTWORK PROCESS can have it's logger
level be configured and also set the logger based on a config value
"rails" or "default". it's a lil clunky but the current implementation
kinda pushed it that way. Yet Another Thing To Come Back To.