fix: dont splat arguments when executing a job nor triggering a pipeline
this is technically a breaking change because we're changing the API of
`Ductwork::Pipeline.trigger` but we're pre-1.0 so, yolo. this should be
the last of ripping out the splatting arguments debacle. will hopefully
revisit this because it would be nice to be able to pass multiple
arguments to `.trigger`. might wait for user research or smthn