fix: correctly create merging steps and jobs for complex pipelines
this is the big enchilada! this fixes both `combine` and `collapse` for
"complex" pipelines. what is a complex pipeline? well for this case it
was a pipeline that expands and divides or divides and expands. the
problem was it was merging _everything_ together not "branches". this is
another big reason to introduce the concept of branches to the system,
but that is going to be a huge change or at least many sequential
changes. either way, this fixes a buncha stuff so that now the
`examples` application runs correctly!!!!!! BIG MILESTONE!!!

the code quality is taking a nosedive. i desperately need to spend some
time refactoring and extracting concepts/classes. thats a "later
problem". time for lunch