chore: let `zeitwerk` autoload models from `lib/models` directory
i like this approach much better as it removes some reliance on rails.
secondly, we're using `zeitwerk` as the code loader so we should let it
load all the code. this does make things a bit more straightforward as
there's not multiple "things" loading code in various ways.

this also paves the way for us to eager load all the code which we will
likely need to do in the `ductwork-pro` gem to properly prepend
the "pro" features