feat!: introduce pipeline Context
pipeline context is basically global state that can be accessed in step
classes and be used to pass data between steps outside of the return
value.this is really useful for sidecar data ie. IDs of other models
that you need for processing. context values are ideally write-once but
there is functionality to overwrite values. this comes with the caveat
that concurrent steps writing to the context can stomp on each other
making the value non-deterministic. honestly if that's happening there
is a smell in the pipeline/step design.

BREAKING CHANGE: because step classes now need to inherit from
`Ductwork::Step`; they are no longer strict POROs